requirements.txt
great_tables[extra]
pandas
polars
python-fasthtml
pytest